improved services and api

+++ src/main/java/com/cub/project/repository/GroupRepository.java
@@ -3,5 +3,5 @@ package com.cub.project.repository;
 import com.cub.project.domain.models.Group;
 import org.springframework.data.repository.CrudRepository;
 
-public interface GroupRepository extends CrudRepository<Group, Integer> {
+public interface GroupRepository extends CrudRepository<Group, Long> {
 }
